Introduce Hi-Unicast Hack
If netCache receives a fileAnnouncement from an unknown host, it's useless for the user, because we don't know the TCP-Port to download the file from.
netCache delays the FileAnnouncement Message and sends a HI unicast to this unknown host, so that he will response with a BYE message.
The FileAnnoucement is processed afterwards.